% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ss_a520_baseline.R
\name{ss_a520_baseline}
\alias{ss_a520_baseline}
\title{Collect baseline 520 absorbance from a list of ECS data}
\usage{
ss_a520_baseline(ecs_data)
}
\arguments{
\item{ecs_data}{A list of ECS data traces}
}
\description{
Given a list of ECS data (ie from read_all_folder) creates a matching list of baseline A520 data
TODO: Re-write to accept a single file. Until then, you can call ss_a520_baseline(list(datum))
}
