% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ss_add_phino.R
\name{ss_add_phino}
\alias{ss_add_phino}
\title{ss_add_phino}
\usage{
ss_add_phino(df, fm, fo)
}
\arguments{
\item{df}{A tibble with NPQt, qL data.}

\item{fm}{darkadapted fm data}

\item{fo}{darkadapted fo data}
}
\description{
A function that adds Phi_NO and NO_qL to a dataset
}
