% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ss_addNPQ.R
\name{ss_addNPQ}
\alias{ss_addNPQ}
\title{Add NPQ and NPQt to a data frame.}
\usage{
ss_addNPQ(df, fm = NA)
}
\arguments{
\item{df}{A data frame with Fm' and Fo' data.}

\item{fm}{The darkadapted fm}
}
\description{
Given a data frame, returns the same data frame with NPQ and NPQt added.
}
