% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ss_addqL.R
\name{ss_addqL}
\alias{ss_addqL}
\title{Add qL to a dataframe with Fm', Fs, and Fo' data}
\usage{
ss_addqL(df)
}
\arguments{
\item{df}{A data frame with Fm',Fs, and Fo' data.}
}
\description{
Given a data frame, returns the same data frame with qL added.
}
