% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ss_bookkeeping.R
\name{ss_bookkeeping}
\alias{ss_bookkeeping}
\title{General usability function for raw data}
\usage{
ss_bookkeeping(dataframe, recalc_delta_a = F, baselineStart = 80,
  baselineEnd = 99)
}
\arguments{
\item{dataframe}{The data to be treated.}

\item{recalc_delta_a}{Boolean. Do you wish to recalculate DeltaA}

\item{baselineStart}{Baseline choices for recalculating DeltaA}

\item{baselineEnd}{Baseline choices for recalculating DeltaA}
}
\description{
From a single trace data table, returns a data table with useful column names.
Optionally recalculates DeltaA if you feel it is necessary.
}
