% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ss_ecs_fit.R
\name{ss_ecs_fit}
\alias{ss_ecs_fit}
\title{Performs fitting of ECS traces}
\usage{
ss_ecs_fit(dataframe, recalc_delta_a = F, graph = F, linFitCount = 5,
  nonlinFitCount = 35, remake = F, baselineStart = 50,
  baselineEnd = 99, abs520 = F, linadj = T)
}
\arguments{
\item{dataframe}{Single trace ECS data point.}

\item{recalc_delta_a}{Do you wish to recalculate DeltaA? Calls ss_bookkeeping}

\item{graph}{Boolean. Recreates the trace with fitting visualized. A list of graphs is returned.}

\item{linFitCount}{Number of points used for linear fitting for vH+.}

\item{nonlinFitCount}{Number of points used for nonlinear fitting for pmf and gH+.}

\item{remake}{Boolean. Improves the detail of recreated graph. Data is untransformed.}

\item{baselineStart}{Baseline location for recalculating deltaA}

\item{baselineEnd}{Baseline location for recalculating DeltaA}

\item{abs520}{Would you like automatic calculation of baseline 520nm absorbance?}

\item{linadj}{Boolean. Attempts to correct for signal drift by straightening out non-DIRK part of the trace}
}
\description{
Given a single trace, generates vH+, gH+, ECSt, and optionally A520 baseline.
Uses minpack.lm for fitting
}
