% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ss_ecs_fit_all.R
\name{ss_ecs_fit_all}
\alias{ss_ecs_fit_all}
\title{Vectorized fitting of ECS data.}
\usage{
ss_ecs_fit_all(ecs_list, recalc_delta_a = F, graph = F,
  linFitCount = 5, nonlinFitCount = 35, remake = F,
  baselineStart = 70, baselineEnd = 99, abs520 = F, linadj = T,
  dirkstart = 100, highWeightVal = 1, highWeightCount = 15)
}
\arguments{
\item{ecs_list}{vector of ecs data to analyze}

\item{recalc_delta_a}{Do you wish to recalculate DeltaA? Calls ss_bookkeeping}

\item{graph}{Boolean. Recreates the trace with fitting visualized. A list of graphs is returned.}

\item{linFitCount}{Number of points used for linear fitting for vH+.}

\item{nonlinFitCount}{Number of points used for nonlinear fitting for pmf and gH+.}

\item{remake}{Boolean. Improves the detail of recreated graph. Data is untransformed.}

\item{baselineStart}{Baseline location for recalculating deltaA}

\item{baselineEnd}{Baseline location for recalculating DeltaA}

\item{abs520}{Would you like automatic calculation of baseline 520nm absorbance?}

\item{linadj}{Boolean. Attempts to correct for signal drift by straightening out non-DIRK part of the trace}

\item{dirkstart}{Integer. The last illuminated data point. Default: 100.}

\item{highWeightVal}{Integer. The amount to weight high-weight points at the start of the dirk.}

\item{highWeightCount}{Integer. The number of points to weight highly at the start of the dirk.}
}
\description{
Functionally applies ecs_fit for a vectorized set of data
Recommended application: ss_read_all_folder %>% ss_splitfun %>% (function(x) x$ecs) %>% ss_ecs_fit_all
}
