% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ss_full_workup.R
\name{ss_full_workup}
\alias{ss_full_workup}
\title{Narrow function that attempts to correlate gas exchange and SharkeySpec data.}
\usage{
ss_full_workup(geLoc, ssLoc, dirklen, satflashlen, timeoffset, graph = F,
  remake = F, baseline = F, recalc = T)
}
\arguments{
\item{geLoc}{location of gas exchange data file.}

\item{ssLoc}{Folder location of Sharkeyspec data}

\item{dirklen}{Length of DIRKS in data points}

\item{satflashlen}{Length of saturating flashes in data points}

\item{timeoffset}{Difference between elapsed time on the li-cor and the sharkeyspec. Subtracted from li-cor elapsed time}

\item{graph}{Boolean. Recreate trace graphs for PS2 and ECS data}

\item{remake}{Boolean. Increase the detail level provided in the graph}

\item{baseline}{Boolean. Provide baseline A520 measurements.}
}
\description{
Requries minpack.lm, reshape2, and readLicorData packages.
}
