% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ss_listslicer.R
\name{ss_listslicer}
\alias{ss_listslicer}
\title{ss_listslicer}
\usage{
ss_listslicer(df, r1, r2)
}
\arguments{
\item{df}{The dataframe to be sliced}

\item{r1}{The first interval}

\item{r2}{The second interval}
}
\description{
Cuts up a df into two lists, based on two intervals provided. I.e., two sets of data, 
one 11 measurements, one 13 measurements, are repeated in a data frame. 
Cuts each of the 11 measurement sets and each of the 13 measurement sets out, 
groups them in a list, and returns the two lists.
}
