% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ss_phi2_calculation.R
\name{ss_phi2_calculation}
\alias{ss_phi2_calculation}
\title{Calculate Phi2 from PS2 traces}
\usage{
ss_phi2_calculation(ps2_data, graph = F, flashstartpoint = 100,
  flashendpoint = 200, spikeomit = 10)
}
\arguments{
\item{ps2_data}{The list of traces to calculate data from}

\item{graph}{Boolean. Will create a graph of the traces with lines drawn for Fm' and Fs. Used to double-check that the data is being automatically calculated correctly}

\item{flashstartpoint}{The data point at which the saturating flash begins. Default: 100}

\item{flashendpoint}{The data point at which the saturating flash ends. Default: 200}

\item{spikeomit}{The script will ignore <<num>> data points after the start of the flash so as not to fit spikes. Default: 10}
}
\description{
Vectorized function. Apply it to a list of traces.
TODO: Allow single traces. Until then use ss_phi2_calculation(list(datum))
}
