% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ss_phi2_calculation_bu.R
\name{ss_phi2_calculation_bu}
\alias{ss_phi2_calculation_bu}
\title{Backup for phi2 data calculation
I don't think I need this any more but I'm keeping it around
Will give it namespace but not export it...}
\usage{
ss_phi2_calculation_bu(ps2_data, graph = F)
}
\description{
Backup for phi2 data calculation
I don't think I need this any more but I'm keeping it around
Will give it namespace but not export it...
}
