% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ss_read.R
\name{ss_read}
\alias{ss_read}
\title{Read a single file of data}
\usage{
ss_read(location)
}
\arguments{
\item{location}{The path of data to use (relative or absolute)}
}
\description{
Mostly for internal use
}
