% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ss_read_all_folder.R
\name{ss_read_all_folder}
\alias{ss_read_all_folder}
\title{Read in and aggregate all files in the folder}
\usage{
ss_read_all_folder(location, ecspattern = "0002", ps2pattern = "0001",
  ps1pattern = "0003", foppattern = "0004")
}
\arguments{
\item{location}{The FOLDER location}

\item{ecspattern}{The default trace number for ECS data}

\item{ps2pattern}{The default trace number for PS2 data}

\item{ps1pattern}{The default trace number for PS1 data}

\item{foppattern}{The default trace number for Fo' data}
}
\description{
Given a file location, reads in all data files and collect them into lists based on file name patterns
The patterns are the numeric appendix on all files based on trace number
}
