% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ss_readType.R
\name{ss_readType}
\alias{ss_readType}
\title{Given a pattern, returns a read-in of all data files matching that pattern}
\usage{
ss_readType(location, pattern)
}
\arguments{
\item{location}{The folder location where data files are}

\item{pattern}{The pattern to match for reading in files}
}
\description{
Given a pattern, returns a read-in of all data files matching that pattern
}
