% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ss_runsetter.R
\name{ss_runsetter}
\alias{ss_runsetter}
\title{A narrow function to read through an appended dataset and attempt to automatically split it up.}
\usage{
ss_runsetter(df, threshhold = 2)
}
\arguments{
\item{df}{The data table (singular) to analyze}

\item{threshhold}{The difference in time that is to be considered a new measurement.}
}
\description{
Use on appended data files where multiple measurements are made in one file. Will create an "ID" column for each individual measurement, as determined by time threshhold.
When the difference in time is greater than the provided threshhold it will decide that it is a new measurement.
}
