% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ss_sorter.R
\name{ss_sorter}
\alias{ss_sorter}
\title{Sort a list of data by time}
\usage{
ss_sorter(a_list)
}
\arguments{
\item{a_list}{The data to be sorted}
}
\description{
I'm certain there's a better way to do this...
}
