% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ss_splitfun.R
\name{ss_splitfun}
\alias{ss_splitfun}
\title{A more rigid application of ss_runsetter}
\usage{
ss_splitfun(df, splitint = 250)
}
\arguments{
\item{df}{The dataframe to split}

\item{splitint}{The length of each individual measurement}
}
\description{
Accepts a dataframe and creates a list of dataframes divided into even chunks based on splitint
Useful for appended homogenous traces
}
