% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ss_timesplitter.R
\name{ss_timesplitter}
\alias{ss_timesplitter}
\title{Carves a dataframe into a list based on time indices.}
\usage{
ss_timesplitter(df, timer, timer2, column = "elapsed")
}
\arguments{
\item{df}{The dataframe to be split.}

\item{timer}{A vector of times to split the chosen column by, start of each interval.}

\item{timer2}{A vector of times to split the chosen column by, end of each interval.}
}
\description{
Given a data frame, carves a list into chunks based on vectors of times.
Functionally vectorizes dplyr::filter
}
