% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ss_ecs_fit.R, R/ss_fit_generic.R
\name{ss_ecs_fit}
\alias{ss_ecs_fit}
\title{Performs fitting of ECS traces}
\usage{
ss_ecs_fit(dataframe, recalc_delta_a = F, graph = F, linFitCount = 5,
  nonlinFitCount = 35, remake = F, baselineStart = 50,
  baselineEnd = 99, abs520 = F, linadj = T, dirkstart = 100)

ss_ecs_fit(dataframe, recalc_delta_a = F, graph = F, linFitCount = 5,
  nonlinFitCount = 35, remake = F, baselineStart = 50,
  baselineEnd = 99, abs520 = F, linadj = T, dirkstart = 100)
}
\arguments{
\item{dataframe}{Single trace ECS data point.}

\item{recalc_delta_a}{Do you wish to recalculate DeltaA? Calls ss_bookkeeping}

\item{graph}{Boolean. Recreates the trace with fitting visualized. A list of graphs is returned.}

\item{linFitCount}{Integer. Number of points used for linear fitting for vH+.}

\item{nonlinFitCount}{Integer. Number of points used for nonlinear fitting for pmf and gH+.}

\item{remake}{Boolean. Improves the detail of recreated graph. Data is untransformed.}

\item{baselineStart}{Integer. Data point startpoint for recalculating deltaA}

\item{baselineEnd}{Integer. Data point endpoint for recalculating DeltaA}

\item{abs520}{Boolean. Would you like automatic calculation of baseline 520nm absorbance?}

\item{linadj}{Boolean. Attempts to correct for signal drift by straightening out non-DIRK part of the trace}

\item{dirkstart}{Integer. The data point count of the last illuminated point.}

\item{dataframe}{Single trace ECS data point.}

\item{recalc_delta_a}{Do you wish to recalculate DeltaA? Calls ss_bookkeeping}

\item{graph}{Boolean. Recreates the trace with fitting visualized. A list of graphs is returned.}

\item{linFitCount}{Integer. Number of points used for linear fitting for vH+.}

\item{nonlinFitCount}{Integer. Number of points used for nonlinear fitting for pmf and gH+.}

\item{remake}{Boolean. Improves the detail of recreated graph. Data is untransformed.}

\item{baselineStart}{Integer. Data point startpoint for recalculating deltaA}

\item{baselineEnd}{Integer. Data point endpoint for recalculating DeltaA}

\item{abs520}{Boolean. Would you like automatic calculation of baseline 520nm absorbance?}

\item{linadj}{Boolean. Attempts to correct for signal drift by straightening out non-DIRK part of the trace}

\item{dirkstart}{Integer. The data point count of the last illuminated point.}
}
\description{
Given a single trace, generates vH+, gH+, ECSt, and optionally A520 baseline.
Uses minpack.lm for fitting

Given a single trace, generates vH+, gH+, ECSt, and optionally A520 baseline.
Uses minpack.lm for fitting
}
