% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ss_fop.R
\name{ss_fop}
\alias{ss_fop}
\title{Attempt to work up Fo' data}
\usage{
ss_fop(ps2_data, graph = F)
}
\arguments{
\item{ps2_data}{A list of PS2 traces}

\item{graph}{Boolean. Reconstitutes a graph with chosen points as lines. Default F. Used to check the accuracy of the program.}
}
\description{
Given a list of PS2 data, generate Fm', fs, phi2, and fo'
Assumes a lot about the shape of your curve.
}
